Outline the 3D character controller for the demo application
