+++ src/quartz/application/Application.cpp
@@ -103,7 +103,7 @@ void quartz::Application::run() {
 
         double frameInterpolationFactor = (frameTimeAccumulator + targetTickTimeDelta) / targetTickTimeDelta;
 
-        currentScene.update(m_renderingContext.getRenderingWindow(), currentFrameTimeDelta, frameInterpolationFactor);
+        currentScene.update(m_renderingContext.getRenderingWindow(), m_inputManager, currentFrameTimeDelta, frameInterpolationFactor);
         m_renderingContext.draw(currentScene);
     }
