+++ src/quartz/scene/camera/Camera.cpp
@@ -124,29 +124,29 @@ quartz::scene::Camera::fixedUpdate(
 
     // ----- update position ----- //
 
-    const float movementSpeedMPS = 2.0f;
-    const float movementSpeedAdjusted = movementSpeedMPS * 0.01666666666;
+    // const float movementSpeedMPS = 2.0f;
+    // const float movementSpeedAdjusted = movementSpeedMPS * 0.01666666666;
 
-    if (inputManager.getKeyDown_w()) {
-        m_worldPosition += movementSpeedAdjusted * currentLookVector;
-    }
-    if (inputManager.getKeyDown_s()) {
-        m_worldPosition -= movementSpeedAdjusted * currentLookVector;
-    }
+    // if (inputManager.getKeyDown_w()) {
+    //     m_worldPosition += movementSpeedAdjusted * currentLookVector;
+    // }
+    // if (inputManager.getKeyDown_s()) {
+    //     m_worldPosition -= movementSpeedAdjusted * currentLookVector;
+    // }
 
-    if (inputManager.getKeyDown_d()) {
-        m_worldPosition += movementSpeedAdjusted * currentRightVector;
-    }
-    if (inputManager.getKeyDown_a()) {
-        m_worldPosition -= movementSpeedAdjusted * currentRightVector;
-    }
+    // if (inputManager.getKeyDown_d()) {
+    //     m_worldPosition += movementSpeedAdjusted * currentRightVector;
+    // }
+    // if (inputManager.getKeyDown_a()) {
+    //     m_worldPosition -= movementSpeedAdjusted * currentRightVector;
+    // }
 
-    if (inputManager.getKeyDown_space()) {
-        m_worldPosition += movementSpeedAdjusted * worldUpVector;
-    }
-    if (inputManager.getKeyDown_shift()) {
-        m_worldPosition -= movementSpeedAdjusted * worldUpVector;
-    }
+    // if (inputManager.getKeyDown_space()) {
+    //     m_worldPosition += movementSpeedAdjusted * worldUpVector;
+    // }
+    // if (inputManager.getKeyDown_shift()) {
+    //     m_worldPosition -= movementSpeedAdjusted * worldUpVector;
+    // }
 }
 
 void
